+++ src/audio/sample_queue.hpp
@@ -13,7 +13,7 @@ public:
     AudioSampleQueue& operator=(AudioSampleQueue&&) = default;
 
     void writeData(const DecodedOpusData& data);
-    void writeData(float* pcm, size_t length);
+    void writeData(const float* pcm, size_t length);
     // contrary to the name, this will erase the samples from this queue after copying them to `dest`
     size_t copyTo(float* dest, size_t samples);
     size_t size() const;